+++ nuxt.config.ts
@@ -1,4 +1,13 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
     ssr: false,
+
+    pages: true,
+
+    generate: {
+        routes: [
+            '/',
+            '/api/calendar'
+        ]
+    }
 })
