+++ package.json
@@ -13,6 +13,8 @@
     "nuxt": "3.0.0"
   },
   "dependencies": {
-    "@commitlint/config-conventional": "^17.8.0"
+    "@commitlint/config-conventional": "^17.8.0",
+    "ical.js": "^1.5.0",
+    "ics": "^3.7.2"
   }
 }
